backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# LangChain Ecosystem - Updated compatible versions
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0

# AI & ML
openai>=1.10.0
anthropic>=0.25.0

# Search & Tools
tavily-python>=0.3.0
google-search-results>=2.4.2
wikipedia>=1.4.0

# Vector Database
chromadb>=0.4.0

# Document Processing
pypdf>=3.17.0

# Configuration & Environment  
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.0.0

# Security & Encryption
cryptography>=41.0.0

# Async Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# Utilities
toposort>=1.10

# Development & Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
